% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyCapture.R
\name{Capture}
\alias{Capture}
\title{Capture}
\usage{
Capture(ElementID, FileName, FileFormat)
}
\arguments{
\item{ElementID}{Pass the ID of the webpage element that needs to be captured. Pass the values 'html' or 'body' to take capture the whole html or body of the webpage.}

\item{FileName}{Filename of the downloaded screenshot file.}

\item{FileFormat}{File format of the downloaded screenshot file. 'jpg' and 'png' are the two supported formats.}
}
\description{
Server part of the package to pass the element ID, file name and format to take and download the screenshot
}
