% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{html2canvas_deps}
\alias{html2canvas_deps}
\title{HTML dependency for vendored html2canvas}
\usage{
html2canvas_deps()
}
\value{
An \code{htmlDependency} object.
}
\description{
Returns an \code{htmltools::htmlDependency} that loads the bundled
html2canvas 1.4.1 library. This is called automatically by
\code{\link{capture_element}} when \code{\link{html2canvas_strategy}} is
used, so you rarely need to call it directly.
}
\details{
Call it manually only if you use \code{\link{capture_plotly}} with a
custom strategy that relies on \code{window.html2canvas}.
}
\examples{
\dontrun{
# Only needed when using html2canvas outside of capture_element()
ui <- fluidPage(
  html2canvas_deps(),
  div(id = "my_div", "Content to capture")
)
}
}
\seealso{
\code{\link{shinycapture_deps}}, \code{\link{capture_element}},
\code{\link{html2canvas_strategy}}
}
