% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyCapture.R
\docType{package}
\name{ShinyCapture}
\alias{ShinyCapture}
\title{ShinyCapture}
\description{
Take easy screenshots of webpage elements in a RShiny Application using Shinyjs package developed by Dean Attali (https://github.com/daattali/shinyjs).
}
\details{
\code{ShinyCapture} can be implemented as follows.

Add the function \code{useShinyCapture} in the UI part of your RShiny web application which will load up the required scripts for taking and downloading screenshots.
The JS can then invoked by using the function \code{Capture} on the server side by passing the element ID, file name of the screenshot and file format.

Click \href{https://sarthi2395.shinyapps.io/ShinyCapture}{here} to see a demo.
}
